+++ src/client/microsoftAuth.js
@@ -11,10 +11,11 @@ async function authenticate (client, options) {
   if (options.authTitle === undefined) {
     options.authTitle = Titles.MinecraftNintendoSwitch
     options.deviceType = 'Nintendo'
+    options.flow = 'live'
   }
 
   const Authflow = new PrismarineAuth(options.username, options.profilesFolder, options, options.onMsaCode)
-  const { token, entitlements, profile } = await Authflow.getMinecraftJavaToken({ fetchProfile: true }).catch(e => {
+  const { token, entitlements, profile, certificates } = await Authflow.getMinecraftJavaToken({ fetchProfile: true, fetchCertificates: !options.disableChatSigning }).catch(e => {
     if (options.password) console.warn('Sign in failed, try removing the password field\n')
     if (e.toString().includes('Not Found')) console.warn(`Please verify that the account ${options.username} owns Minecraft\n`)
     throw e
@@ -32,8 +33,10 @@ async function authenticate (client, options) {
     selectedProfile: profile,
     availableProfile: [profile]
   }
+  Object.assign(client, certificates)
   client.session = session
   client.username = profile.name
+
   options.accessToken = token
   client.emit('session', session)
   options.connect(client)