+++ src/client/setProtocol.js
@@ -15,7 +15,7 @@ module.exports = function (client, options) {
     function next () {
       let taggedHost = options.host
       if (client.tagHost) taggedHost += client.tagHost
-      if (client.fakeHost) taggedHost = options.fakeHost
+      if (options.fakeHost) taggedHost = options.fakeHost
 
       client.write('set_protocol', {
         protocolVersion: options.protocolVersion,
@@ -25,7 +25,14 @@ module.exports = function (client, options) {
       })
       client.state = states.LOGIN
       client.write('login_start', {
-        username: client.username
+        username: client.username,
+        signature: client.profileKeys
+          ? {
+              timestamp: BigInt(client.profileKeys.expiresOn.getTime()), // should probably be called "expireTime"
+              publicKey: client.profileKeys.publicDER,
+              signature: client.profileKeys.signature
+            }
+          : null
       })
     }
   }