+++ test/serverTest.js
@@ -3,77 +3,100 @@
 const mc = require('../')
 const assert = require('power-assert')
 const { once } = require('events')
+const nbt = require('prismarine-nbt')
+const applyClientHelpers = require('./common/clientHelpers')
 
 const { getPort } = require('./common/util')
 
-const w = {
-  piglin_safe: {
-    type: 'byte',
-    value: 0
-  },
-  natural: {
-    type: 'byte',
-    value: 1
-  },
-  ambient_light: {
-    type: 'float',
-    value: 0
-  },
-  infiniburn: {
-    type: 'string',
-    value: 'minecraft:infiniburn_overworld'
-  },
-  respawn_anchor_works: {
-    type: 'byte',
-    value: 0
-  },
-  has_skylight: {
-    type: 'byte',
-    value: 1
-  },
-  bed_works: {
-    type: 'byte',
-    value: 1
-  },
-  has_raids: {
-    type: 'byte',
-    value: 1
-  },
-  name: {
-    type: 'string',
-    value: 'minecraft:overworld'
-  },
-  logical_height: {
-    type: 'int',
-    value: 256
-  },
-  shrunk: {
-    type: 'byte',
-    value: 0
-  },
-  ultrawarm: {
-    type: 'byte',
-    value: 0
-  },
-  has_ceiling: {
-    type: 'byte',
-    value: 0
-  }
-}
+const w = nbt.comp({
+  piglin_safe: nbt.byte(0),
+  natural: nbt.byte(1),
+  ambient_light: nbt.float(0),
+  infiniburn: nbt.string('minecraft:infiniburn_overworld'),
+  respawn_anchor_works: nbt.byte(0),
+  has_skylight: nbt.byte(1),
+  bed_works: nbt.byte(1),
+  has_raids: nbt.byte(1),
+  name: nbt.string('minecraft:overworld'),
+  logical_height: nbt.int(256),
+  shrunk: nbt.byte(0),
+  ultrawarm: nbt.byte(0),
+  has_ceiling: nbt.byte(0)
+})
 
 for (const supportedVersion of mc.supportedVersions) {
   let PORT
   const mcData = require('minecraft-data')(supportedVersion)
   const version = mcData.version
 
-  describe('mc-server ' + version.minecraftVersion, function () {
+  const loginPacket = (client, server) => {
+    return {
+      // 1.7
+      entityId: client.id,
+      gameMode: 1,
+      dimension: (version.version >= 735 ? mcData.loginPacket.dimension : 0),
+      difficulty: 2,
+      maxPlayers: server.maxPlayers,
+      levelType: 'default',
+      // 1.8
+      reducedDebugInfo: (version.version >= 735 ? false : 0),
+      // 1.14
+      // removes `difficulty`
+      viewDistance: 10,
+      // 1.15
+      hashedSeed: [0, 0],
+      enableRespawnScreen: true,
+      // 1.16
+      // removed levelType
+      previousGameMode: version.version >= 755 ? 0 : 255,
+      worldNames: ['minecraft:overworld'],
+      dimensionCodec: version.version >= 755 ? mcData.loginPacket.dimensionCodec : (version.version >= 735 ? mcData.loginPacket.dimension : { name: '', type: 'compound', value: { dimension: { type: 'list', value: { type: 'compound', value: [w] } } } }),
+      worldName: 'minecraft:overworld',
+      isDebug: false,
+      isFlat: false,
+      // 1.16.2
+      isHardcore: false,
+      // 1.18
+      simulationDistance: 10,
+      // 1.19
+      // removed `dimension`
+      // removed `dimensionCodec`
+      registryCodec: {
+        "type": "compound",
+        "name": "",
+        "value": {}
+      },
+      worldType: "minecraft:overworld",
+      death: undefined
+      // more to be added
+    }
+  }
 
+  function sendBroadcastMessage(server, clients, message, sender) {
+    if (mcData.supportFeature('signedChat')) {
+      server.writeToClients(clients, 'player_chat', {
+        signedChatContent: '',
+        unsignedChatContent: JSON.stringify({ text: message }),
+        type: 0,  
+        senderUuid: 'd3527a0b-bc03-45d5-a878-2aafdd8c8a43', // random
+        senderName: JSON.stringify({ text: sender }),
+        senderTeam: undefined,
+        timestamp: Date.now(),
+        salt: 0n,
+        signature: Buffer.alloc(0)
+      })
+    } else {
+      server.writeToClients(clients, 'chat', { message: JSON.stringify({ text: message }), position: 0, sender: sender || '0' })
+    }
+  }
+
+  describe('mc-server ' + version.minecraftVersion, function () {
+    this.timeout(5000)
     this.beforeAll(async function() {
       PORT = await getPort()
       console.log(`Using port for tests: ${PORT}`)
     })
 
-    this.timeout(5000)
     it('starts listening and shuts down cleanly', function (done) {
       const server = mc.createServer({
         'online-mode': false,
@@ -90,6 +113,7 @@ for (const supportedVersion of mc.supportedVersions) {
         done()
       })
     })
+
     it('kicks clients that do not log in', function (done) {
       const server = mc.createServer({
         'online-mode': false,
@@ -105,14 +129,10 @@ for (const supportedVersion of mc.supportedVersions) {
           server.close()
         })
       })
-      server.on('close', function () {
-        resolve()
-      })
+      server.on('close', resolve)
       server.on('listening', function () {
         const client = new mc.Client(false, version.minecraftVersion)
-        client.on('end', function () {
-          resolve()
-        })
+        client.on('end', resolve)
         client.connect(PORT, '127.0.0.1')
       })
 
@@ -121,6 +141,7 @@ for (const supportedVersion of mc.supportedVersions) {
         if (count <= 0) done()
       }
     })
+
     it('kicks clients that do not send keepalive packets', function (done) {
       const server = mc.createServer({
         'online-mode': false,
@@ -136,9 +157,7 @@ for (const supportedVersion of mc.supportedVersions) {
           server.close()
         })
       })
-      server.on('close', function () {
-        resolve()
-      })
+      server.on('close', resolve)
       server.on('listening', function () {
         const client = mc.createClient({
           username: 'superpants',
@@ -147,15 +166,14 @@ for (const supportedVersion of mc.supportedVersions) {
           keepAlive: false,
           version: version.minecraftVersion
         })
-        client.on('end', function () {
-          resolve()
-        })
+        client.on('end', resolve)
       })
       function resolve () {
         count -= 1
         if (count <= 0) done()
       }
     })
+
     it('responds to ping requests', function (done) {
       const chatMotd = { // Generated with prismarine-chat MessageBuilder on version 1.16 may change in the future
         extra: [{ color: 'red', text: 'Red text' }],
@@ -191,7 +209,7 @@ for (const supportedVersion of mc.supportedVersions) {
               online: 0,
               sample: []
             },
-            description: { 
+            description: {
               extra: [ { color: 'red', text: 'Red text' } ],
               bold: true,
               text: 'Example chat mesasge'
@@ -202,6 +220,7 @@ for (const supportedVersion of mc.supportedVersions) {
       })
       server.on('close', done)
     })
+
     it('responds with chatMessage motd\'s', function (done) {
       const server = mc.createServer({
         'online-mode': false,
@@ -237,6 +256,7 @@ for (const supportedVersion of mc.supportedVersions) {
       })
       server.on('close', done)
     })
+
     it('clients can be changed by beforeLogin', function (done) {
       const notchUUID = '069a79f4-44e9-4726-a5be-fca90e38aaf5'
       const server = mc.createServer({
@@ -263,12 +283,16 @@ for (const supportedVersion of mc.supportedVersions) {
       })
       server.on('close', done)
     })
+
     it('clients can log in and chat', function (done) {
       const server = mc.createServer({
         'online-mode': false,
         version: version.minecraftVersion,
         port: PORT
       })
+      const broadcast = (message, exclude) => sendBroadcastMessage(server,
+        Object.values(server.clients).filter(client => client !== exclude), message)
+
       const username = ['player1', 'player2']
       let index = 0
       server.on('login', function (client) {
@@ -279,86 +303,49 @@ for (const supportedVersion of mc.supportedVersions) {
           broadcast(client.username + ' left the game.', client)
           if (client.username === 'player2') server.close()
         })
-        const loginPacket = {
-          entityId: client.id,
-          levelType: 'default',
-          gameMode: 1,
-          previousGameMode: version.version >= 755 ? 0 : 255,
-          worldNames: ['minecraft:overworld'],
-          dimensionCodec: version.version >= 755 ? mcData.loginPacket.dimensionCodec : (version.version >= 735 ? mcData.loginPacket.dimension : { name: '', type: 'compound', value: { dimension: { type: 'list', value: { type: 'compound', value: [w] } } } }),
-          dimension: (version.version >= 735 ? mcData.loginPacket.dimension : 0),
-          worldName: 'minecraft:overworld',
-          hashedSeed: [0, 0],
-          difficulty: 2,
-          maxPlayers: server.maxPlayers,
-          reducedDebugInfo: (version.version >= 735 ? false : 0),
-          enableRespawnScreen: true
-        }
-        if (version.version >= 735) { // 1.16x
-          loginPacket.isDebug = false
-          loginPacket.isFlat = false
-          loginPacket.isHardcore = false
-          loginPacket.viewDistance = 10
-          delete loginPacket.levelType
-          delete loginPacket.difficulty
-        }
-        client.write('login', loginPacket)
-        client.on('chat', function (packet) {
-          const message = '<' + client.username + '>' + ' ' + packet.message
-          broadcast(message)
-        })
+        client.write('login', loginPacket(client, server))
+
+        const handleChat = (packet) => broadcast(`<${client.username}> ${packet.message}`)
+        client.on('chat', handleChat)
+        client.on('chat_message', handleChat)
       })
       server.on('close', done)
+
       server.on('listening', function () {
-        const player1 = mc.createClient({
+        const player1 = applyClientHelpers(mc.createClient({
           username: 'player1',
           host: '127.0.0.1',
           version: version.minecraftVersion,
           port: PORT
-        })
-        player1.on('login', function (packet) {
+        }))
+        player1.on('login', async function (packet) {
           assert.strictEqual(packet.gameMode, 1)
-          player1.once('chat', function (packet) {
-            assert.strictEqual(packet.message, '{"text":"player2 joined the game."}')
-            player1.once('chat', function (packet) {
-              assert.strictEqual(packet.message, '{"text":"<player2> hi"}')
-              player2.once('chat', fn)
-              function fn (packet) {
-                if (/<player2>/.test(packet.message)) {
-                  player2.once('chat', fn)
-                  return
-                }
-                assert.strictEqual(packet.message, '{"text":"<player1> hello"}')
-                player1.once('chat', function (packet) {
-                  assert.strictEqual(packet.message, '{"text":"player2 left the game."}')
-                  player1.end()
-                })
-                player2.end()
-              }
-
-              player1.write('chat', { message: 'hello' })
-            })
-            player2.write('chat', { message: 'hi' })
-          })
-          const player2 = mc.createClient({
+          const player2 = applyClientHelpers(mc.createClient({
             username: 'player2',
             host: '127.0.0.1',
             version: version.minecraftVersion,
             port: PORT
-          })
+          }))
+
+          const p1Join = await player1.nextMessage('player2')
+          assert.strictEqual(p1Join, '{"text":"player2 joined the game."}')
+
+          player2.chat('hi')
+          const p2hi = await player1.nextMessage('player2')
+          assert.strictEqual(p2hi, '{"text":"<player2> hi"}')
+
+          player1.chat('hello')
+          const p1hello = await player2.nextMessage('player1')
+          assert.strictEqual(p1hello, '{"text":"<player1> hello"}')
+
+          player2.end()
+          const p2leaving = await player1.nextMessage('player2')
+          assert.strictEqual(p2leaving, '{"text":"player2 left the game."}')
+          player1.end()
         })
       })
-
-      function broadcast (message, exclude) {
-        let client
-        for (const clientId in server.clients) {
-          if (server.clients[clientId] === undefined) continue
-
-          client = server.clients[clientId]
-          if (client !== exclude) client.write('chat', { message: JSON.stringify({ text: message }), position: 0, sender: '0' })
-        }
-      }
     })
+
     it('kicks clients when invalid credentials', function (done) {
       this.timeout(10000)
       const server = mc.createServer({
@@ -372,9 +359,7 @@ for (const supportedVersion of mc.supportedVersions) {
           server.close()
         })
       })
-      server.on('close', function () {
-        resolve()
-      })
+      server.on('close', resolve)
       server.on('listening', function () {
         resolve()
         const client = mc.createClient({
@@ -383,15 +368,14 @@ for (const supportedVersion of mc.supportedVersions) {
           version: version.minecraftVersion,
           port: PORT
         })
-        client.on('end', function () {
-          resolve()
-        })
+        client.on('end', resolve)
       })
       function resolve () {
         count -= 1
         if (count <= 0) done()
       }
     })
+
     it('gives correct reason for kicking clients when shutting down', function (done) {
       const server = mc.createServer({
         'online-mode': false,
@@ -404,34 +388,9 @@ for (const supportedVersion of mc.supportedVersions) {
           assert.strictEqual(reason, 'ServerShutdown')
           resolve()
         })
-        const loginPacket = {
-          entityId: client.id,
-          levelType: 'default',
-          gameMode: 1,
-          previousGameMode: version.version >= 755 ? 0 : 255,
-          worldNames: ['minecraft:overworld'],
-          dimensionCodec: version.version >= 755 ? mcData.loginPacket.dimensionCodec : (version.version >= 735 ? mcData.loginPacket.dimension : { name: '', type: 'compound', value: { dimension: { type: 'list', value: { type: 'compound', value: [w] } } } }),
-          dimension: (version.version >= 735 ? mcData.loginPacket.dimension : 0),
-          worldName: 'minecraft:overworld',
-          hashedSeed: [0, 0],
-          difficulty: 2,
-          maxPlayers: server.maxPlayers,
-          reducedDebugInfo: (version.version >= 735 ? false : 0),
-          enableRespawnScreen: true
-        }
-        if (version.version >= 735) { // 1.16x
-          loginPacket.isDebug = false
-          loginPacket.isFlat = false
-          loginPacket.isHardcore = false
-          loginPacket.viewDistance = 10
-          delete loginPacket.levelType
-          delete loginPacket.difficulty
-        }
-        client.write('login', loginPacket)
+        client.write('login', loginPacket(client, server))
       })
-      server.on('close', function () {
-        resolve()
-      })
+      server.on('close', resolve)
       server.on('listening', function () {
         const client = mc.createClient({
           username: 'lalalal',
@@ -448,6 +407,7 @@ for (const supportedVersion of mc.supportedVersions) {
         if (count <= 0) done()
       }
     })
+
     it('encodes chat packet once and send it to two clients', function (done) {
       const server = mc.createServer({
         'online-mode': false,
@@ -455,51 +415,31 @@ for (const supportedVersion of mc.supportedVersions) {
         port: PORT
       })
       server.on('login', function (client) {
-        const loginPacket = {
-          entityId: client.id,
-          levelType: 'default',
-          gameMode: 1,
-          previousGameMode: version.version >= 755 ? 0 : 255,
-          worldNames: ['minecraft:overworld'],
-          dimensionCodec: version.version >= 755 ? mcData.loginPacket.dimensionCodec : (version.version >= 735 ? mcData.loginPacket.dimension : { name: '', type: 'compound', value: { dimension: { type: 'list', value: { type: 'compound', value: [w] } } } }),
-          dimension: (version.version >= 735 ? mcData.loginPacket.dimension : 0),
-          worldName: 'minecraft:overworld',
-          hashedSeed: [0, 0],
-          difficulty: 2,
-          maxPlayers: server.maxPlayers,
-          reducedDebugInfo: (version.version >= 735 ? false : 0),
-          enableRespawnScreen: true
-        }
-        if (version.version >= 735) { // 1.16x
-          loginPacket.isDebug = false
-          loginPacket.isFlat = false
-          loginPacket.isHardcore = false
-          loginPacket.viewDistance = 10
-          delete loginPacket.levelType
-          delete loginPacket.difficulty
-        }
-        client.write('login', loginPacket)
+        client.write('login', loginPacket(client, server))
       })
       server.on('close', done)
       server.on('listening', async function () {
-        const player1 = mc.createClient({
+        const player1 = applyClientHelpers(mc.createClient({
           username: 'player1',
           host: '127.0.0.1',
           version: version.minecraftVersion,
           port: PORT
-        })
-        const player2 = mc.createClient({
+        }))
+        const player2 = applyClientHelpers(mc.createClient({
           username: 'player2',
           host: '127.0.0.1',
           version: version.minecraftVersion,
           port: PORT
-        })
-        await Promise.all([once(player1, 'login'), once(player2, 'login')])        
-        server.writeToClients(Object.values(server.clients), 'chat', { message: JSON.stringify({ text: 'A message from the server.' }), position: 1, sender: '00000000-0000-0000-0000-000000000000' })
-        
-        let results = await Promise.all([ once(player1, 'chat'), once(player2, 'chat') ])
-        results.forEach(res => assert.strictEqual(res[0].message, '{"text":"A message from the server."}'))
-        
+        }))
+        await Promise.all([once(player1, 'login'), once(player2, 'login')])
+
+        sendBroadcastMessage(server, Object.values(server.clients), 'A message from the server.')
+
+        let results = await Promise.all([player1.nextMessage(), player2.nextMessage()])
+        for (const msg of results) {
+          assert.strictEqual(msg, '{"text":"A message from the server."}')
+        }
+
         player1.end()
         player2.end()
         await Promise.all([once(player1, 'end'), once(player2, 'end')])